requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==1.9.0
torchvision==0.10.0
numpy==1.21.2
Pillow==8.3.2
pyyaml==5.4.1
opencv-python==4.5.3.56
easyocr==1.6.2 